Correction to the Russian telephone number formatting.
Previously when entering a Russian telephone number there were not enough digits and the formatting was incorrect and missing the brackets for the area code.
Now when entering a Russian telephone number it is the correct format for 98% of commonly used numbers.

Further information can be found at, https://properrussian.com/2011/06/russian-phone-numbers.html